Macro files for the analysis
Open root in the same directory of rootlogon.c. It will load the macro.cpp file. By default it will search for the root file specified in the ifilename TString variable defined at the beginning of the macro.cpp code. Please change filename if the ROOT output file from the MC simulation is placed in another path.